// src/app/interfaces/sign.ts
import { Elements } from './elements';
export type PossibleSigns = 'aries' | 'taurus' | 'gemini' | 'cancer' | 'leo' | 'virgo' | 'libra' | 'scorpio' | 'sagittarius' | 'capricorn' | 'aquarius' | 'pisces';
export interface Sign {
  collectionId: string; // "ggpwskw1cs9fls9"
  collectionName: "signs"
  created: string; // "2022-12-30 12:53:09.680Z"
  detriment: string; // "ko86e1r2qrpka3c"
  element: Elements;
  exaltation: string; // "jeftfqh40q41wun"
  fall: string; // "6s85ntb6qmhc5ms"
  id: string; // "2zrtospisqo7qfh"
  name: PossibleSigns; // "aries"
  quality: string; // "cardinal"
  ruler: string; // "yf2dyccbcnt2zr3"
  updated: string; // "2022-12-30 12:55:27.304Z"
  color: string;
}